backend/requirements.txt
# --- Core Django ---
Django==5.0.6
djangorestframework==3.15.1
django-cors-headers==4.3.1
django-filter==24.2

# --- Auth ---
djangorestframework-simplejwt==5.3.1

# --- Utils ---
python-dotenv==1.0.1
Pillow==10.3.0           # gestion images (photos matériel)

# --- Dev/test ---
pytest==8.2.0
pytest-django==4.8.0

# --- IA (utile dès la Phase 5, on l'installera plus tard) ---
# transformers==4.41.0
# peft==0.11.1
# bitsandbytes==0.43.1
# accelerate==0.30.0
# fastapi==0.111.0
# uvicorn==0.29.0
